Enable scripts to use two db servers.
The species list has been expanded to include some plants,
but these are stored on a different database server than
the vertebrates. Rather than complicate things with logic
to work out which server to use, just load both into the
registry - use versioned urls rather than a registry file
for ease of configuration (not only would you have to work
out where to store file and ensure the service can find it,
it's tricky to enable flexibility in setting the release version).

Since we now process the registry urls in the 'Submit' module,
it's easy to use them to retrieve the assembly from the core db,
so that it does not need to be hard-coded in the config.
This also functions as a handy check on the availability
of the the species in the given servers - if not there, the
pipeline will fail right away, without doing further jobs that
would leave the data in a partial state that requires rollbacks.